Refactored FullText Queries
